New Updated Textures and Bool hasKey

+++ Game/src/GridMaker.cpp
@@ -50,7 +50,7 @@ int GridMaker::index_tile(char c) {
 }
 
 //Get tile char from grid
-char GridMaker::get_tile(sf::Vector2i position) {
+char GridMaker::get_tile(sf::Vector2f position) {
 	int x = position.x / 16;
 	int y = position.y / 16;
 
@@ -58,7 +58,7 @@ char GridMaker::get_tile(sf::Vector2i position) {
 }
 
 //Get tile phase from grid
-TileType GridMaker::check_tile(sf::Vector2i position) {
+TileType GridMaker::check_tile(sf::Vector2f position) {
 	char c = get_tile(position);
 
 	switch(c) {
